requirements.txt
flask
wheel
python-dotenv
exchangelib